+++ data/provider/src/main/java/tech/pegasys/teku/api/blobselector/BlobSidecarSelectorFactory.java
@@ -178,12 +178,32 @@ public class BlobSidecarSelectorFactory extends AbstractSelectorFactory<BlobSide
 
   private SafeFuture<Optional<List<BlobSidecar>>> getBlobSidecars(
       final SlotAndBlockRoot slotAndBlockRoot, final List<UInt64> indices) {
-    return client.getBlobSidecars(slotAndBlockRoot, indices).thenApply(Optional::of);
+    return client
+        .getBlobSidecars(slotAndBlockRoot, indices)
+        .thenCompose(
+            blobSidecars -> {
+              if (blobSidecars.isEmpty()) {
+                // attempt retrieving from archive (when enabled)
+                return client.getArchivedBlobSidecars(slotAndBlockRoot, indices);
+              }
+              return SafeFuture.completedFuture(blobSidecars);
+            })
+        .thenApply(Optional::of);
   }
 
   private SafeFuture<Optional<List<BlobSidecar>>> getBlobSidecars(
       final UInt64 slot, final List<UInt64> indices) {
-    return client.getBlobSidecars(slot, indices).thenApply(Optional::of);
+    return client
+        .getBlobSidecars(slot, indices)
+        .thenCompose(
+            blobSidecars -> {
+              if (blobSidecars.isEmpty()) {
+                // attempt retrieving from archive (when enabled)
+                return client.getArchivedBlobSidecars(slot, indices);
+              }
+              return SafeFuture.completedFuture(blobSidecars);
+            })
+        .thenApply(Optional::of);
   }
 
   private Optional<BlobSidecarsAndMetaData> addMetaData(