+++ storage/api/src/main/java/tech/pegasys/teku/storage/api/StorageQueryChannel.java
@@ -56,7 +56,7 @@ public interface StorageQueryChannel extends ChannelInterface {
   SafeFuture<Optional<SignedBlockAndState>> getHotBlockAndStateByBlockRoot(Bytes32 blockRoot);
 
   SafeFuture<Optional<StateAndBlockSummary>> getHotStateAndBlockSummaryByBlockRoot(
-      final Bytes32 blockRoot);
+      Bytes32 blockRoot);
 
   /**
    * Returns "hot" blocks - the latest finalized block or blocks that descend from the latest
@@ -109,4 +109,9 @@ public interface StorageQueryChannel extends ChannelInterface {
 
   SafeFuture<List<SlotAndBlockRootAndBlobIndex>> getBlobSidecarKeys(
       SlotAndBlockRoot slotAndBlockRoot);
+
+  // Methods for retrieving archived blobs (when enabled)
+  SafeFuture<List<BlobSidecar>> getArchivedBlobSidecars(SlotAndBlockRoot slotAndBlockRoot);
+
+  SafeFuture<List<BlobSidecar>> getArchivedBlobSidecars(UInt64 slot);
 }