+++ teku/src/main/java/tech/pegasys/teku/cli/options/Eth2NetworkOptions.java
@@ -159,6 +159,14 @@ public class Eth2NetworkOptions {
   private UInt64 electraForkEpoch;
 
   @Option(
+      names = {"--Xnetwork-fulu-fork-epoch"},
+      hidden = true,
+      paramLabel = "<epoch>",
+      description = "Override the Fulu fork activation epoch.",
+      arity = "1")
+  private UInt64 fuluForkEpoch;
+
+  @Option(
       names = {"--Xnetwork-total-terminal-difficulty-override"},
       hidden = true,
       paramLabel = "<uint256>",
@@ -306,6 +314,11 @@ public class Eth2NetworkOptions {
               + ".io/get-started/connect).");
     }
 
+    if (initialState != null && checkpointSyncUrl != null) {
+      throw new InvalidConfigurationException(
+          "Both --initial-state and --checkpoint-sync-url are provided. Please specify only one.");
+    }
+
     builder.applyNetworkDefaults(network);
     if (startupTargetPeerCount != null) {
       builder.startupTargetPeerCount(startupTargetPeerCount);
@@ -339,6 +352,9 @@ public class Eth2NetworkOptions {
     }
     if (electraForkEpoch != null) {
       builder.electraForkEpoch(electraForkEpoch);
+    }
+    if (fuluForkEpoch != null) {
+      builder.fuluForkEpoch(fuluForkEpoch);
     }
     if (totalTerminalDifficultyOverride != null) {
       builder.totalTerminalDifficultyOverride(totalTerminalDifficultyOverride);